Add `MetricsGenerator` class
