+++ esgf_metrics/parse.py
@@ -1,6 +1,5 @@
 """Parse module for parsing ESGF Apache access logs"""
 import os
-from collections import defaultdict
 from pathlib import Path
 from typing import DefaultDict, List, Optional, TypedDict
 
@@ -92,27 +91,6 @@ class LogParser:
         # DataFrame for storing parsed log lines from the available log files.
         self.df_log_request: Optional[pd.DataFrame] = self._parse_log_requests()
 
-        # DataFrame of fiscal monthly metrics by project.
-        self.df_monthly_metrics: pd.DataFrame = pd.DataFrame()
-        self.df_fiscal_metrics: pd.DataFrame = pd.DataFrame()
-
-        # Dictionary with the key being the project and the value being
-        # a sub-dictionary of facet reports. The key of the
-        # sub-dictionary is the facet name and the value is a DataFrame storing
-        # the report.
-        """
-        Example:
-        {
-            "E3SM and E3SM CMIP6": {"realm": pd.DataFrame()},
-            "E3SM": {"time_frequency": pd.DataFrame()},
-            "E3SM CMIP6": {"activity": pd.DataFrame(), "variable_id": pd.DataFrame()},
-        }
-        """
-
-        self.fiscal_facet_metrics: LogParser.FiscalFacetMetrics = defaultdict(
-            lambda: defaultdict(pd.DataFrame)
-        )
-
     def to_sql(self):
         """
         Inserts `self.df_log_file` and `self.df_log_request` to the `log_file`
@@ -142,33 +120,6 @@ class LogParser:
         )
 
         logger.info("Database insertion completed")
-
-    def get_metrics(self):
-        """Generates metrics for successful requests by project and facets.
-        Raises
-        ------
-        ValueError
-            If logs have not been parsed yet.
-        """
-        if self.df_log_request is None:
-            raise ValueError(
-                "Logs have not been parsed yet, call `parse_logs()` first."
-            )
-        logger.info("Generating monthly metrics.")
-        self.df_monthly_metrics = self._get_monthly_metrics()
-        self.df_fiscal_metrics = self._get_fiscal_metrics(
-            self.df_monthly_metrics.copy()
-        )
-
-        for project, facets in AVAILABLE_FACETS.items():
-            for facet in facets.keys():
-                # TODO: We might want to store this in the dictionary for plotting.
-                df_monthly_metrics = self._get_monthly_metrics(facet)
-                df_fy_metrics = self._get_fiscal_metrics(
-                    df_monthly_metrics.copy(), facet
-                )
-                df_fy_metrics.loc[df_fy_metrics.project == project]
-                self.fiscal_facet_metrics[project][facet] = df_fy_metrics
 
     def _parse_log_files(self) -> Optional[pd.DataFrame]:
         """Parses metadata from the access logs.
@@ -454,11 +405,13 @@ class LogParser:
 
         # FIXME: This is suboptimal, it should be done after all log lines
         # are already in the DataFrame.
-        for facets in AVAILABLE_FACETS.values():
-            for facet, option in facets.items():
-                facet_value = None
+        for facet, options in AVAILABLE_FACETS.values():
+            facet_value = None
+            for option in options:
                 if option in dataset_id:
                     facet_value = option
+                    break
+
             log_line[facet] = facet_value  # type: ignore
 
         return log_line
@@ -506,6 +459,27 @@ class LogParser:
 
         return df
 
+    def _add_fiscal_date_cols(self, df: pd.DataFrame) -> pd.DataFrame:
+        """Adds fiscal date columns based on the calendar date columns.
+        This method resamples the monthly metrics DataFrame to extract the
+        fiscal year, quarter, and month using the calendar year and month.
+        Parameters
+        ----------
+        df : pd.DataFrame
+            The monthly metrics DataFrame.
+        Returns
+        -------
+        pd.DataFrame
+            The monthly metrics DataFrame with fiscal date columns.
+        """
+        df["fiscal_year_quarter"] = df.apply(
+            lambda row: row.year_month.asfreq("Q-JUN"), axis=1
+        )
+        df["fiscal_year"] = df.fiscal_year_quarter.dt.strftime("%F").astype("int")
+        df["fiscal_month"] = df.apply(lambda row: E3SM_CY_TO_FY_MAP[row.month], axis=1)
+
+        return df
+
     def _cast_period_cols_to_str(self, df: pd.DataFrame) -> pd.DataFrame:
         """Cast the pd.Period dtype to str.
 
@@ -528,151 +502,3 @@ class LogParser:
             df["fiscal_year_quarter"] = df["fiscal_year_quarter"].astype(str)
 
         return df
-
-    def _get_monthly_metrics(self, facet: Optional[str] = None) -> pd.DataFrame:
-        # FIXME: Use SQL queries to fetch from the database.
-        df = self.df_log_request.copy()  # type: ignore
-        agg_cols = ["project", "calendar_year_month", "calendar_year", "calendar_month"]
-
-        if facet is not None:
-            agg_cols.insert(1, facet)
-            df[facet] = df[facet].fillna(value="N/A")
-
-        # Get total requests (includes all request types) and total GET requests
-        # by month.
-        df_reqs = df.value_counts(subset=agg_cols).reset_index(name="total_requests")
-
-        df_log_get_reqs = df[df.status_code.str.contains("200|206")]
-        df_get_reqs = df_log_get_reqs.value_counts(subset=agg_cols).reset_index(
-            name="total_get_requests"
-        )
-
-        df_reqs = pd.merge(df_reqs, df_get_reqs, on=agg_cols)
-
-        # Get total data downloaded by month (only successful requests).
-        df_data_size = (
-            df_log_get_reqs.groupby(by=agg_cols).agg({"mb": "sum"}).reset_index()
-        )
-        df_data_size["total_gb"] = df_data_size.mb.div(1024)
-        df_data_size = df_data_size.drop(columns="mb")
-
-        # Merge DataFrames into a single DataFrame
-        df_final = pd.merge(df_reqs, df_data_size, on=agg_cols)
-        df_final = df_final.sort_values(by=agg_cols)
-
-        return df_final
-
-    def _get_fiscal_metrics(
-        self, df_monthly: pd.DataFrame, facet: Optional[str] = None
-    ):
-        """Generates monthly metrics by project and facet (optionally).
-        Parameters
-        ----------
-        df_monthly : pd.DataFrame
-            The monthly metrics DataFrame.
-        facet : Optional[str], optional
-            The facet column used for grouping in addition to the main
-            aggregation columns, by default None.
-        Returns
-        -------
-        pd.DataFrame
-            A DataFrame of monthly metrics.
-        """
-        # Get fiscal monthly cumulative sums for requests and size of data
-        # download.
-        df = self._add_fiscal_date_cols(df_monthly)
-        df = self._get_fiscal_monthly_cumsums(df, facet)
-        df = self._reorder_columns(df, facet)
-
-        return df
-
-    def _add_fiscal_date_cols(self, df: pd.DataFrame) -> pd.DataFrame:
-        """Adds fiscal date columns based on the calendar date columns.
-        This method resamples the monthly metrics DataFrame to extract the
-        fiscal year, quarter, and month using the calendar year and month.
-        Parameters
-        ----------
-        df : pd.DataFrame
-            The monthly metrics DataFrame.
-        Returns
-        -------
-        pd.DataFrame
-            The monthly metrics DataFrame with fiscal date columns.
-        """
-        df["fiscal_year_quarter"] = df.apply(
-            lambda row: row.year_month.asfreq("Q-JUN"), axis=1
-        )
-        df["fiscal_year"] = df.fiscal_year_quarter.dt.strftime("%F").astype("int")
-        df["fiscal_month"] = df.apply(lambda row: E3SM_CY_TO_FY_MAP[row.month], axis=1)
-
-        return df
-
-    def _get_fiscal_monthly_cumsums(
-        self, df: pd.DataFrame, facet: Optional[str]
-    ) -> pd.DataFrame:
-        """Adds cumulative sum columns for requests and data downloads (GB).
-
-        Parameters
-        ----------
-        df_monthly : pd.DataFrame
-            The DataFrame of monthly metrics
-        facet : Optional[str]
-            The facet column used for grouping in addition to the main
-            aggregation columns.
-
-        Returns
-        -------
-        pd.DataFrame
-            The DataFrame of monthly metrics with cumulative sums.
-        """
-        gb_cols = ["project", "fiscal_year"]
-        if facet is not None:
-            gb_cols.append(facet)
-
-        df["cumulative_requests"] = df.groupby(gb_cols)["total_requests"].cumsum()
-        df["cumulative_get_requests"] = df.groupby(gb_cols)[
-            "total_get_requests"
-        ].cumsum()
-        df["cumulative_gb"] = df.groupby(gb_cols)["total_gb"].cumsum()
-
-        return df
-
-    def _reorder_columns(self, df, facet: Optional[str]):
-        """Reorders columns for a DataFrame of aggregated metrics.
-
-        In several class methods, DataFrame columns are appended to the end of
-        the DataFrame, which makes the order of columns not succinct.
-
-        Parameters
-        ----------
-        df : pd.DataFrame
-            A DataFrame of aggregated metrics with the original column order.
-        facet : Optional[str]
-            The facet column used for grouping, in addition to the main
-            aggregation columns.
-
-        Returns
-        -------
-        pd.DataFrame
-            A DataFrame of aggregated metrics with reordered columns.
-        """
-        columns = [
-            "project",
-            "fiscal_year_quarter",
-            "fiscal_year",
-            "fiscal_month",
-            "fiscal_quarter",
-            "year_month",
-            "year",
-            "month",
-            "total_requests",
-            "cumulative_requests",
-            "total_get_requests",
-            "cumulative_get_requests",
-            "total_gb",
-            "cumulative_gb",
-        ]
-        if facet is not None:
-            columns.insert(1, facet)
-
-        return df[columns]
